Includes the import flag for nib stylus plugin
Nib needs two things: setting up on stylus core ( --use option ), which
is already used and import the library to the stylus files.

The import can be done on command line with --import option. With this
we could leave imported nib library without specifying on each stylus
file.

Also this feature depends on
https://github.com/LearnBoost/stylus/pull/797

on the stylus files.